requirements.txt
streamlit
python-dotenv
empire-chain
discord.py>=2.3.2
openai>=1.0.0
python-jose[cryptography]
aiohttp
APScheduler
redis
sqlalchemy
pydantic
python-multipart
PyNaCl>=1.5.0
supabase